Задача №6: Отработка арифметических операций Изменён тип данных на число с плавающей точкой для корректного расчёта среднего арифметического числа

+++ src/Task_6.java
@@ -3,12 +3,12 @@ public class Task_6 {
     public static void main (String[] args) {
         Scanner in = new Scanner(System.in);
         System.out.print("Введите x: ");
-        double x = in.nextInt();
+        int x = in.nextInt();
         System.out.print("Введите y: ");
-        double y = in.nextInt();
+        int y = in.nextInt();
         System.out.print("Введите z: ");
-        double z = in.nextInt();
-        double avr = (x+y+z)/3;
+        int z = in.nextInt();
+        double avr = (double) (x+y+z)/3;
         System.out.println("Среднее арифметическое чисел x, y, z: " + avr);
         double i = avr/2;
         if (i > 3)
